refactor react-quiz: reduce prop-drilling by using QuizContext
