+++ src/components/Footer.js
@@ -1,5 +1,3 @@
-import NextButton from "./NextButton";
-
 function Footer({ children }) {
   return <footer>{children}</footer>;
 }
