+++ src/components/Progress.js
@@ -1,4 +1,8 @@
-function Progress({ index, numQuestions, score, maxScore, answer }) {
+import { useQuiz } from "../contexts/QuizContext";
+
+function Progress() {
+  const { index, numQuestions, score, maxScore, answer } = useQuiz();
+
   return (
     <>
       <progress
