+++ src/components/Question.js
@@ -1,14 +1,13 @@
+import { useQuiz } from "../contexts/QuizContext";
 import Options from "./Options";
 
-function Question({ question, dispatch, answer }) {
+function Question() {
+  const { questions, index } = useQuiz();
+
   return (
     <div>
-      <h4>{question.question}</h4>
-      <Options
-        question={question}
-        dispatch={dispatch}
-        answer={answer}
-      />
+      <h4>{questions[index].question}</h4>
+      <Options question={questions[index]} />
     </div>
   );
 }
