+++ src/components/Timer.js
@@ -1,8 +1,12 @@
 import { useEffect } from "react";
+import { useQuiz } from "../contexts/QuizContext";
 
-function Timer({ dispatch, secondsRemaining }) {
+function Timer() {
+  const { dispatch, secondsRemaining } = useQuiz();
+
   const mins = Math.floor(secondsRemaining / 60);
   const secs = secondsRemaining % 60;
+
   useEffect(
     function () {
       const id = setInterval(function () {
@@ -13,6 +17,7 @@ function Timer({ dispatch, secondsRemaining }) {
     },
     [dispatch]
   );
+
   return (
     <div className="timer">
       {mins < 10 ? 0 : ""}
